+++ scripts/blocks.js
@@ -1,2 +1,3 @@
 // import './blocks/rebates_page_info/index.js';
 import './blocks/multi-query/index.js';
+import './blocks/query-conditional-group/index.js';